requirements.txt
transformers
Pillow
pypdf
jq
streamlit
boto3
awscli
pandas-datareader
langchain-experimental
bs4
setuptools
PyPDF2
pypdf
langchain==0.2.7
langchain-community==0.2.7
python-dotenv
botocore